refactor(strategy): bound LLMStrategy with CPU fallback
Cap the Ollama call timeout (default 4s, was 15s) and replace the
hardcoded Bid(2,3)/CHALLENGE fallbacks with a delegating CPUStrategy.
Observation events are now forwarded to the fallback so its opponent
model stays in sync.

+++ src/llm_client.py
@@ -11,7 +11,7 @@ def _ollama_url() -> str:
     return os.environ.get('OLLAMA_URL', _DEFAULT_OLLAMA_URL)
 
 
-def query_llm(model: str, prompt: str, timeout: int = 15, temperature: float = 0.7) -> str | None:
+def query_llm(model: str, prompt: str, timeout: float = 15, temperature: float = 0.7) -> str | None:
     try:
         response = requests.post(
             _ollama_url(),
